refactor(storybook): .vaadin-button
* theme as options knob
* rename files cxl-vaadin-button -> vaadin-button
