+++ packages/cxl-lumo-styles/scss/global.scss
@@ -1,5 +1,5 @@
 /** Integrations. */
-@use 'integrations/cxl-vaadin-button.scss';
+@use 'integrations/vaadin-button.scss';
 
 html {
   // Variables.
